demo/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.8.1
certifi==2023.7.22
charset-normalizer==3.2.0
cytoolz==0.12.2
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
fastecdsa==2.3.0
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.24.2
py-solc-x==1.1.1
pycryptodome==3.18.0
pyunormalize==15.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.0
rusty-rlp==0.2.1
semantic-version==2.10.0
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.4
web3==6.9.0
websockets==11.0.3
yarl==1.9.2
